Fix imported website URLs
Some URLs in the source feed are missing their http(s):// scheme, and this causes issues with the link field.

This technique uses a 'Negative Lookahead' (https://www.regular-expressions.info/lookaround.html) to see if the scheme is missing and add one if so.

It's a little crazy that it works!