Add drupal/sqlite337 to support systems with sqlite < 3.45
platform.sh uses Debian-based images and are stuck on 3.40.x. This is unlikely to change soon.

The reason for the core requirement of >=3.45 is for JSONB support, but nothing in core requires it at the moment.

See https://www.drupal.org/project/drupal/issues/3346338 for all the gory details.

We could have just stayed running on 10.4 for the next couple of years, but I really wanted to test D11. We may come to regret the decision.

¯\_(ツ)_/¯